contrib: use 64-bit element offsets in SkipLayerNorm CUDA write index
Support large tensors in SkipLayerNormalization by computing the global output
element offset in 64-bit. For big rows * hidden_size (rows = batch * sequence
length), which can exceed 2^31 elements on high-memory GPUs, the previous 32-bit
row offset arithmetic could not represent the index.

- skip_layer_norm_impl.cu: compute the per-block offset/index in int64 in both
  SkipLayerNormKernel and SkipLayerNormKernelSmall.
- layer_norm.cuh: widen the global write offset/index of SimplifiedLayerNorm,
  LayerNormSmall and SimplifiedLayerNormSmall (the helpers used only by
  SkipLayerNorm) to int64_t. The gamma/beta/bias indices (ld, i,
  threadIdx.x * ILP) stay int since they are bounded by hidden_size.

LayerNorm() (used by EmbedLayerNorm) is widened separately in the
EmbedLayerNorm write-index change. Numerics are unchanged for normal sizes.

+++ onnxruntime/contrib_ops/cuda/bert/skip_layer_norm_impl.cu
@@ -75,7 +75,7 @@ __global__ void SkipLayerNormKernel(
     T* output, T* sum_output, const T* input, const T* skip, const T* bias, const T* gamma, const T* beta,
     float epsilon, const int ld, int skip_size) {
   const float reverse_ld = 1.f / ld;
-  const int offset = blockIdx.x * ld;
+  const int64_t offset = static_cast<int64_t>(blockIdx.x) * ld;
   const bool has_bias = (bias != nullptr);
 
   // Reduce sum of x and x^2, and the results are divided by ld.
@@ -84,7 +84,7 @@ __global__ void SkipLayerNormKernel(
   cub::KeyValuePair<float, float> thread_data(0.f, 0.f);
 
   for (int i = threadIdx.x; i < ld; i += TPB) {
-    const int idx = offset + i;
+    const int64_t idx = offset + i;
 
     T val = input[idx];
     if (has_bias) {
@@ -116,7 +116,7 @@ __global__ void SkipLayerNormKernelSmall(
     T* output, T* sum_output, const T* input, const T* skip, const T* bias, const T* gamma, const T* beta,
     float epsilon, int ld, int skip_size) {
   const float rld = 1.f / ld;
-  const int idx = blockIdx.x * ld + threadIdx.x * ILP;
+  const int64_t idx = static_cast<int64_t>(blockIdx.x) * ld + threadIdx.x * ILP;
 
   using VecT = aligned_vector<T, ILP>;
   T sum_v[ILP];
